Split the internal blueprint into multiple view files
