+++ core/admin/mailu/internal/views/__init__.py
@@ -1,0 +1,3 @@
+__all__ = [
+    'auth', 'postfix', 'dovecot', 'fetch'
+]
